# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tender API",
    "version": "0.1.0",
    "description": "Any-chain crypto checkout. Buyers pay from any chain; merchants settle in one asset on Monad.\n\nMerchant routes need **Authorize → `tk_live_…`** (create one with `npm run merchant:create`). Public routes are what the buyer checkout calls. Money is always a decimal **string**; the wire is **snake_case**."
  },
  "components": {
    "securitySchemes": {
      "merchantKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Merchant API key (tk_live_…)"
      },
      "platformKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dashboard platform key (tp_…). On /v1 routes, also send `X-Tender-Merchant: mer_…`. Server-side only."
      }
    }
  },
...
  "paths": {
...
    "/v1/transfers/wallet": {
      "get": {
        "tags": [
          "Transfers"
        ],
        "summary": "Can this merchant send money out, and what does their wallet hold?",
        "description": "Reads the settlement wallet's on-chain balance of its settlement asset. `can_send` is false, with a `reason`, when the wallet is not verified, the asset cannot be sent through Tender, or sending is switched off.",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "address": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "asset": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "balance": {
                      "anyOf": [
                        {
                          "type": "string",
                          "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "can_send": {
                      "type": "boolean"
                    },
                    "reason": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "can_send"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/transfers/{id}/submit": {
      "post": {
        "tags": [
          "Transfers"
        ],
        "summary": "Step 2: submit the wallet's signatures; Tender relays them and pays the gas",
        "description": "Each signature must come from the settlement wallet. A split goes through Multicall3 as ONE transaction, so it lands whole or not at all. The response is SUBMITTED; it becomes CONFIRMED only when the chain shows a successful receipt and every authorization is spent.",
        "security": [
          {
            "merchantKey": []
          },
          {
            "platformKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "signatures": {
                    "minItems": 1,
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^0x[0-9a-fA-F]{130}$"
                    }
                  }
                },
                "required": [
                  "signatures"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "kind": {
                      "type": "string",
                      "enum": [
                        "PAYOUT",
                        "REFUND",
                        "SPLIT"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "AWAITING_SIGNATURE",
                        "SUBMITTED",
                        "CONFIRMED",
                        "FAILED",
                        "EXPIRED"
                      ]
                    },
                    "asset": {
                      "type": "string"
                    },
                    "from": {
                      "type": "string"
                    },
                    "total_amount": {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                    },
                    "payment_id": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "note": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "tx_hash": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "failure_reason": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "lines": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "to": {
                            "type": "string"
                          },
                          "amount": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          }
                        },
                        "required": [
                          "to",
                          "amount"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "submitted_at": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "confirmed_at": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "authorizations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "index": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "typed_data": {
                            "type": "object",
                            "properties": {
                              "domain": {
                                "type": "object",
                                "properties": {
                                  "name": {
                                    "type": "string"
                                  },
                                  "version": {
                                    "type": "string"
                                  },
                                  "chainId": {
                                    "type": "number"
                                  },
                                  "verifyingContract": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "name",
                                  "version",
                                  "chainId",
                                  "verifyingContract"
                                ],
                                "additionalProperties": false
                              },
                              "types": {
                                "type": "object",
                                "propertyNames": {
                                  "type": "string"
                                },
                                "additionalProperties": {
                                  "type": "array",
                                  "items": {
                                    "type": "object",
                                    "properties": {
                                      "name": {
                                        "type": "string"
                                      },
                                      "type": {
                                        "type": "string"
                                      }
                                    },
                                    "required": [
                                      "name",
                                      "type"
                                    ],
                                    "additionalProperties": false
                                  }
                                }
                              },
                              "primaryType": {
                                "type": "string",
                                "const": "TransferWithAuthorization"
                              },
                              "message": {
                                "type": "object",
                                "propertyNames": {
                                  "type": "string"
                                },
                                "additionalProperties": {
                                  "type": "string"
                                }
                              }
                            },
                            "required": [
                              "domain",
                              "types",
                              "primaryType",
                              "message"
                            ],
                            "additionalProperties": false
                          }
                        },
                        "required": [
                          "index",
                          "typed_data"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "id",
                    "kind",
                    "status",
                    "asset",
                    "from",
                    "total_amount",
                    "lines",
                    "created_at",
                    "expires_at"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — `fields` names each bad field",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflict with the current state",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "410": {
            "description": "Expired — start again",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Aurora could not be reached; nothing was written, retry",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Not ready yet (or, on the assistant, the model is unavailable)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}